se separa la parte de grupos de la pantalla principal
