+++ Lifora/backoffice.cs
@@ -98,181 +98,29 @@ namespace Lifora
                 textBoxCambiarEmail.Text = seleccion.Cells[4].Value?.ToString();
                 textBoxFechaDeNacimiento.Text = seleccion.Cells[5].Value?.ToString();
                 id = seleccion.Cells[0].Value?.ToString();
-                dataGridViewGrupos.DataSource = ControladorGrupos.ListarGrupos();
-                dataGridViewEventos.DataSource = ControladorEventos.ListarEventos();
-
-                if (dataGridViewGrupos.DataSource != null)
-                    (dataGridViewGrupos.DataSource as DataTable).DefaultView.RowFilter = string.Format("Convert(idCuenta, 'System.String') LIKE '%{0}%'", id);
-                if (dataGridViewEventos.SelectedRows.Count > 0)
-                    (dataGridViewEventos.DataSource as DataTable).DefaultView.RowFilter = string.Format("id_cuenta LIKE '%{0}%'", id);
+                
             }
         }
         
-        private void BtnCrearEvento_Click(object sender, EventArgs e)
-        {
-            CrearEventoBackoffice ceb = new CrearEventoBackoffice();
-            ceb.Show();
-        }
-        private void dataGridViewEventos_SelectionChanged(object sender, EventArgs e)
-        {
-            if (dataGridViewEventos.SelectedRows.Count > 0)
-            {
-                DataGridViewRow seleccion = dataGridViewEventos.SelectedRows[0];
-                textBoxNuevoNombreEvento.Text = seleccion.Cells[1].Value?.ToString();
-                textBoxNuevaInfoEvento.Text = seleccion.Cells[2].Value?.ToString();
-                textBoxNuevoLugarEvento.Text = seleccion.Cells[3].Value?.ToString();
-                textBoxNuevaFechaEvento.Text = seleccion.Cells[4].Value?.ToString();
-            }
-        }
-        private void BtnBloquearEvento_click(object sender, EventArgs e)
-        {
-            DialogResult pregunta = MessageBox.Show("Bloquear este Evento?", "Estas seguro?", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-            if (pregunta == DialogResult.Yes)
-            {
-                if (dataGridViewEventos.SelectedRows.Count > 0)
-                {
-                    DataGridViewRow seleccion = dataGridViewEventos.SelectedRows[0];
-                    int columna = 0;
-                    var CellValue = seleccion.Cells[columna].Value;
-                    string Id = CellValue.ToString();
-                    int id_evento = int.Parse(Id);
-                    ControladorEventos.DeshabilitarEvento(id_evento);
-                    dataGridViewInfoUser.DataSource = ControladorCuentaUsuario.Listar();
-                }
-                if (dataGridViewInfoUser.SelectedRows.Count == 0)
-                {
-                    MessageBox.Show("Debes seleccionar un evento");
-                }
-            }
-            if (pregunta == DialogResult.No)
-            {
-                MessageBox.Show("No se ah bloqueado el evento");
-            }
-        }
-        private void BtnDesbloquearEvento_click(object sender, EventArgs e)
-        {
-            DialogResult pregunta = MessageBox.Show("Desbloquear este Evento?", "Estas seguro?", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-            if (pregunta == DialogResult.Yes)
-            {
-                if (dataGridViewEventos.SelectedRows.Count > 0)
-                {
-                    DataGridViewRow seleccion = dataGridViewEventos.SelectedRows[0];
-                    int columna = 0;
-                    var CellValue = seleccion.Cells[columna].Value;
-                    string Id = CellValue.ToString();
-                    int id_evento = int.Parse(Id);
-                    ControladorEventos.HabilitarEvento(id_evento);
-                    dataGridViewInfoUser.DataSource = ControladorCuentaUsuario.Listar();
-                }
-                if (dataGridViewInfoUser.SelectedRows.Count == 0)
-                {
-                    MessageBox.Show("Debes seleccionar un evento");
-                }
-            }
-            if (pregunta == DialogResult.No)
-            {
-                MessageBox.Show("No se ah Habilitado el evento");
-            }
-        }
-        private void BtnModificarEvento_Click_1(object sender, EventArgs e)
-        {
-            DialogResult pregunta = MessageBox.Show("Aplicar cambios?", "Confirmar", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-
-            if (pregunta == DialogResult.Yes)
-            {
-                DataGridViewRow seleccion = dataGridViewEventos.SelectedRows[0];
-                int columna = 0;
-                var CellValue = seleccion.Cells[columna].Value;
-                string id_evento = CellValue.ToString();
-                ControladorEventos.ModificarEventoBackoffice(id_evento, textBoxNuevoNombreEvento.Text, textBoxNuevaInfoEvento.Text, textBoxNuevoLugarEvento.Text, textBoxNuevaFechaEvento.Text);
-                MessageBox.Show("Cambios realizados con exito");
-            }
-            if (pregunta == DialogResult.No)
-            {
-                MessageBox.Show("No se han realizado los cambios");
-            }
-            dataGridViewInfoUser.DataSource = ControladorCuentaUsuario.Listar();
-        }    
-        private void btnCrearGrupo_Click(object sender, EventArgs e)
-        {
-            CrearGrupoBackoffice cgb = new CrearGrupoBackoffice();
-            cgb.Show();
-        }
-        private void btnModificarGrupo_Click(object sender, EventArgs e)
-        {
-
-            DialogResult pregunta = MessageBox.Show("Aplicar cambios?", "Confirmar", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-            if (pregunta != DialogResult.Yes)
-            {
-                MessageBox.Show("No se han realizado los cambios");
-                return;
-            }
-            DataGridViewRow seleccion = dataGridViewGrupos.SelectedRows[0];
-            int columna = 0;
-            var CellValue = seleccion.Cells[columna].Value;
-            int idGrupo = Int32.Parse(CellValue.ToString());
-            string nombre = textBoxNombreGrupo.Text;
-            string descripcion = textBoxIDescripcionGrupo.Text;
-            ControladorGrupos.ModificarGrupo(idGrupo, nombre, descripcion);
-            dataGridViewGrupos.DataSource = ControladorGrupos.ListarGrupos();
-        }
-        private void btnBloquearGrupo_Click(object sender, EventArgs e)
-        {
-            DialogResult pregunta = MessageBox.Show("Bloquear este Grupo?", "Estas seguro?", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-            if (pregunta == DialogResult.Yes)
-            {
-                if (dataGridViewGrupos.SelectedRows.Count > 0)
-                {
-                    DataGridViewRow seleccion = dataGridViewGrupos.SelectedRows[0];
-                    int columna = 0;
-                    var CellValue = seleccion.Cells[columna].Value;
-                    string Id = CellValue.ToString();
-                    int idGrupo = int.Parse(Id);
-                    ControladorGrupos.BloquearGrupo(idGrupo);
-                    dataGridViewGrupos.DataSource = ControladorGrupos.ListarGrupos();
-                }
-                if (dataGridViewGrupos.SelectedRows.Count == 0)
-                {
-                    MessageBox.Show("Debes seleccionar un Grupo");
-                }
-            }
-            if (pregunta == DialogResult.No)
-            {
-                MessageBox.Show("No se ah bloqueado el Grupos");
-            }
-            dataGridViewGrupos.DataSource = ControladorGrupos.ListarGrupos();
-        }
-        private void btnHabilitarGrupo_Click(object sender, EventArgs e)
-        {
-            DialogResult pregunta = MessageBox.Show("Desbloquear este Grupo?", "Estas seguro?", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-            if (pregunta == DialogResult.Yes)
-            {
-                if (dataGridViewGrupos.SelectedRows.Count > 0)
-                {
-                    DataGridViewRow seleccion = dataGridViewGrupos.SelectedRows[0];
-                    int columna = 0;
-                    var CellValue = seleccion.Cells[columna].Value;
-                    string Id = CellValue.ToString();
-                    int idGrupo = int.Parse(Id);
-                    ControladorGrupos.HabilitarGrupo(idGrupo);
-                    dataGridViewGrupos.DataSource = ControladorGrupos.ListarGrupos();
-                }
-                if (dataGridViewGrupos.SelectedRows.Count == 0)
-                {
-                    MessageBox.Show("Debes seleccionar un Grupo");
-                }
-            }
-            if (pregunta == DialogResult.No)
-            {
-                MessageBox.Show("No se ah bloqueado el Grupo");
-            }
-            dataGridViewGrupos.DataSource = ControladorGrupos.ListarGrupos();
-        }
+        
+        
 
         private void button1_Click(object sender, EventArgs e)
         {
             VisualPost post = new VisualPost();
             post.Show();
+        }
+
+        private void button2_Click(object sender, EventArgs e)
+        {
+            VisualEventos veb = new VisualEventos();
+            veb.Show();
+        }
+
+        private void button3_Click(object sender, EventArgs e)
+        {
+            VisualGrupos vgb = new VisualGrupos();
+            vgb.Show();
         }
     }
 }
